low_level/automation/pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "automation"
version = "0.1.0"
description = "Wayland automation library with screenshot and global shortcuts"
requires-python = ">=3.8"
dependencies = ["numpy", "pillow"]

[tool.maturin]
features = ["pyo3/extension-module"]
